requirements.txt
altgraph==0.17.4
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
duckduckgo_search==8.0.2
et_xmlfile==2.0.0
h11==0.16.0
idna==3.10
lxml==5.4.0
macholib==1.16.3
modulegraph==0.19.6
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
primp==0.15.0
py2app==0.28.8
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.32.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
